move entry files and remove empty directory
