+++ license.js
